requirements.txt
azure-cognitiveservices-vision-computervision==0.9.1
Flask==3.1.2
flask_wtf==1.2.2
msrest==0.7.1
python-dotenv==1.2.1
Requests==2.32.5
Werkzeug==3.1.5
WTForms==3.2.1
